Moved MainActivity to OldMainActivity and NewMainActivity to MainActivity
