+++ app/src/main/java/com/kylemsguy/tcasmobile/LoginActivity.java
@@ -7,7 +7,6 @@ import android.app.AlertDialog;
 import android.content.ContentResolver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.PackageInfo;
 import android.database.Cursor;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
@@ -33,7 +32,6 @@ import android.widget.TextView;
 import com.kylemsguy.tcasmobile.apiwrapper.LoginResponse;
 import com.kylemsguy.tcasmobile.backend.SessionManager;
 import com.kylemsguy.tcasmobile.tasks.GetLoggedInTask;
-import com.kylemsguy.tcasmobile.tasks.LogoutTask;
 
 import java.net.CookieStore;
 import java.net.HttpCookie;
@@ -282,7 +280,7 @@ public class LoginActivity extends AppCompatActivity implements GetLoggedInTask.
             if (saveData.isChecked())
                 saveUserData(mUsername, mPassword);
             PrefUtils.saveToPrefs(this, PrefUtils.PREF_LOGGED_IN_KEY, mDisplayUsername);
-            Intent intent = new Intent(this, NewMainActivity.class);
+            Intent intent = new Intent(this, MainActivity.class);
             // TODO temp while we are using both the Mobile and Web APIs
             if (mDisplayUsername == null)
                 mDisplayUsername = mUsername;
