+++ app/src/main/res/layout/content_new_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
 
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.kylemsguy.tcasmobile.NewMainActivity"
+    tools:context=".MainActivity"
     tools:showIn="@layout/app_bar_new_main">
     <!--android:paddingBottom="@dimen/activity_vertical_margin"
        android:paddingLeft="@dimen/activity_horizontal_margin"
