+++ app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.kylemsguy.tcasmobile.MainActivity$PlaceholderFragment" >
+    tools:context=".OldMainActivity$PlaceholderFragment">
 
     <TextView
         android:id="@+id/section_label"
