+++ app/src/main/res/menu/main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.kylemsguy.tcasmobile.MainActivity">
+    tools:context=".OldMainActivity">
 
     <item
         android:id="@+id/action_logout"
